Mocked out the NotifyService
