+++ src/main/resources/development.application.conf
@@ -24,4 +24,5 @@ api {
 mockConfig {
   mockCompanySearch = true
   mockCompanyAuth = true
+  mockNotify = true
 }
