+++ src/main/scala/Module.scala
@@ -21,14 +21,15 @@ import config.{AppConfig, MockConfig}
 import play.api.libs.concurrent.AkkaGuiceSupport
 import play.api.{Configuration, Environment, Logger}
 import services.companiesHouse.{CompaniesHouseAuth, CompaniesHouseSearch}
-import services.mocks.{MockCompanyAuth, MockCompanySearch}
-import services.{CompanyAuthService, CompanySearchService, SessionCleaner}
+import services.mocks.{MockCompanyAuth, MockCompanySearch, MockNotify}
+import services.notify.NotifyServiceImpl
+import services.{CompanyAuthService, CompanySearchService, NotifyService, SessionCleaner}
 import slicks.modules.DB
 
 class Module(environment: Environment, configuration: Configuration) extends AbstractModule with AkkaGuiceSupport {
   override def configure(): Unit = {
 
-    val mockConfig = new AppConfig(configuration).config.mockConfig.getOrElse(MockConfig(None, None))
+    val mockConfig = new AppConfig(configuration).config.mockConfig.getOrElse(MockConfig.empty)
 
     val searchImpl = if (mockConfig.mockCompanySearch.getOrElse(false)) {
       Logger.debug("Wiring in Company Search Mock")
@@ -41,6 +42,12 @@ class Module(environment: Environment, configuration: Configuration) extends Abs
       classOf[MockCompanyAuth]
     } else classOf[CompaniesHouseAuth]
     bind(classOf[CompanyAuthService]).to(authImpl)
+
+    val notifyImpl = if (mockConfig.mockNotify.getOrElse(false)) {
+      Logger.debug("Wiring in Notify Mock")
+      classOf[MockNotify]
+    } else classOf[NotifyServiceImpl]
+    bind(classOf[NotifyService]).to(notifyImpl)
 
     bind(classOf[DB]).asEagerSingleton()
     bindActor[ConfirmationActor]("confirmation-actor")
