+++ src/main/scala/config/Config.scala
@@ -39,7 +39,11 @@ case class ClientConfig(id: String, secret: String)
 
 case class GoogleAnalytics(code: Option[String])
 
-case class MockConfig(mockCompanySearch: Option[Boolean], mockCompanyAuth: Option[Boolean])
+case class MockConfig(mockCompanySearch: Option[Boolean], mockCompanyAuth: Option[Boolean], mockNotify: Option[Boolean])
+
+object MockConfig {
+  val empty = MockConfig(None, None, None)
+}
 
 case class Config(
                    companiesHouse: CompaniesHouseConfig,
